More perf hacking with string interner (#1311)

+++ src/backend/xilinx/axi_address_space.rs
@@ -216,7 +216,7 @@ impl AddressSpace {
                 branch.add_seq(v::Sequential::new_nonblk_assign(
                     slice(
                         data_variable,
-                        self.data_width as u64,
+                        self.data_width,
                         &meaning.address_range,
                     ),
                     self.slice(meaning),
@@ -226,7 +226,7 @@ impl AddressSpace {
 
             if end < 32 {
                 branch.add_seq(v::Sequential::new_nonblk_assign(
-                    slice(data_variable, self.data_width as u64, &(end..32)),
+                    slice(data_variable, self.data_width, &(end..32)),
                     v::Expr::new_int(0),
                 ));
             }
@@ -286,7 +286,7 @@ impl AddressSpace {
             ));
             else_br.add_seq(v::Sequential::new_nonblk_assign(
                 self.slice(meaning),
-                slice(data, self.data_width as u64, &meaning.address_range),
+                slice(data, self.data_width, &meaning.address_range),
             ));
             if let Some(name) = &meaning.flags.clear_on_handshake {
                 let mut clear_if = v::SequentialIfElse::new(name.as_str());
